refactor: generateCalendarDates 함수를 utils/date.ts 하위로 추출
